plexdvrapi: depend on python to allow build
On commit 26ca38ddb the enigma2 removed because it wasn't a real dependency.
Although plugin requires the python developement to be present and on pyro
without expicit depend it fails to compile.

Depend on python in order to build the plugin, although @OpenVIX should update plugin
and remove depends to python. There are only runtime depends to enigma2 and python.